[Quality] Removed Global Warning Suppressions (-bugprone-* etc) (#1823)
[ROCm/MIOpen commit: d2c2e0e7a05a058b12865efd2da3819acf3c0030]

+++ src/ocl/dropoutocl.cpp
@@ -255,8 +255,8 @@ void DropoutDescriptor::DropoutForward(const Handle& handle,
     size_t RD_BLCK    = /* (in_len[4] % 4 == 0) ? 4 : */ (in_len[2] % 2 == 0) ? 2 : 1;
     size_t total_work = (in_len[4] / RD_BLCK) * in_len[3] * in_len[2] * in_len[1] * in_len[0];
 
-    size_t max_wk_grp =
-        use_mask ? MAX_WORKITEM_NUM : std::min(size_t(MAX_PRNG_STATE), handle.GetImage3dMaxWidth());
+    size_t max_wk_grp = use_mask ? size_t(MAX_WORKITEM_NUM)
+                                 : std::min(size_t(MAX_PRNG_STATE), handle.GetImage3dMaxWidth());
     size_t wk_grp_num =
         std::min(max_wk_grp / 256,
                  ((in_len[4] * in_len[3] * in_len[2] * in_len[1] * in_len[0] + 255) / 256));
@@ -458,8 +458,8 @@ void DropoutDescriptor::DropoutBackward(const Handle& handle,
     size_t RD_BLCK    = /* (in_len[4] % 4 == 0) ? 4 : */ (in_len[2] % 2 == 0) ? 2 : 1;
     size_t total_work = (in_len[4] / RD_BLCK) * in_len[3] * in_len[2] * in_len[1] * in_len[0];
 
-    size_t max_wk_grp =
-        use_prng ? std::min(size_t(MAX_PRNG_STATE), handle.GetImage3dMaxWidth()) : MAX_WORKITEM_NUM;
+    size_t max_wk_grp = use_prng ? std::min(size_t(MAX_PRNG_STATE), handle.GetImage3dMaxWidth())
+                                 : size_t(MAX_WORKITEM_NUM);
     size_t wk_grp_num =
         std::min(max_wk_grp / 256,
                  ((in_len[4] * in_len[3] * in_len[2] * in_len[1] * in_len[0] + 255) / 256));
