+++ src/solver/conv_hip_implicit_gemm_bwd_v4r1_xdlops.cpp
@@ -409,7 +409,7 @@ bool PerformanceImplicitGemmBwdDataV4R1Xdlops::IsReallyValid(
     std::size_t lds_size      = 0;
     std::tie(lds_size, valid) = CalculateLdsNumberOfByte(problem);
 
-    return (valid and lds_size <= 64 * 1024);
+    return (valid and lds_size <= static_cast<std::size_t>(64) * 1024);
 }
 
 bool PerformanceImplicitGemmBwdDataV4R1Xdlops::IsFastToBeUsedForTuning(
